+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## python-markdown2 2.4.10 (not yet released)
 
 - [pull #520] Allow more relative links in safe mode (issue #517)
+- [pull #522] Add `middle-word-em` extra
 
 
 ## python-markdown2 2.4.9